requirements.txt
atproto>=0.0.40
Pillow>=10.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
